add tests

+++ firmware/include/PID.h
@@ -6,16 +6,16 @@
 
 
 typedef struct{
-  double  kp;         // the value of the P coefficient
-  double  ki;         // the value of the I coefficient
-  double  kd;         // the value of the D coefficient
+  float  kp;         // the value of the P coefficient
+  float  ki;         // the value of the I coefficient
+  float  kd;         // the value of the D coefficient
 }coef_t;
 
 typedef struct{
-  double  P;          // P component
-  double  prev_P;     // prev P component
-  double  I;          // I component
-  double  D;          // D component
+  float  P;          // P component
+  float  prev_P;     // prev P component
+  float  I;          // I component
+  float  D;          // D component
 }error_type_t;
 
 typedef struct{
@@ -23,7 +23,7 @@ typedef struct{
 	coef_t coef;
 }PID_t;
 
-double PID_out(PID_t* reg, double dt);
+float PID_out(PID_t* reg, float dt);
 void PID_reset(PID_t* reg);
 void PID_err_calc (error_type_t *err_reg, float angle, float current_angle);
