+++ firmware/src/PID.c
@@ -4,10 +4,10 @@
  * @brief PID function
  * @brief recieve time step and PID struct
  */
-double PID_out(PID_t* reg, double dt)
+float PID_out(PID_t* reg, float dt)
 {
 
-	double control = 0;
+  float control = 0;
 
 	reg->error.I += reg->error.P*dt;
 	reg->error.D = (reg->error.P - reg->error.prev_P)/dt;
